+++ src/components/Dialog/Dialog.component.tsx
@@ -34,7 +34,7 @@ const defaultDialogProps: DefaultDialogProps = {
  *
  * Controlled by a parent `DialogProvider`.
  *
- * @version 0.0.3
+ * @version 0.0.4
  *
  * @param {DialogProps & TestIdProps} props - the properties for the Dialog component
  * @returns {Nullable<JSX.Element>} the rendered dialog or null if closed
@@ -52,9 +52,6 @@ const Dialog = (props: DialogProps & TestIdProps): Nullable<JSX.Element> => {
   const isOpenRef = useRef<boolean>(open);
 
   useEffect(() => {
-    if (isRegistered) {
-      return unregister;
-    }
     register(open);
     return unregister;
     // eslint-disable-next-line react-hooks/exhaustive-deps